requirements.txt
torch>=1.13.0
torch-geometric>=2.2.0
numpy>=1.20.0
pyyaml>=6.0
pandas>=1.3.0
matplotlib>=3.5.0
networkx>=2.7.0
tqdm>=4.62.0
scikit-learn>=1.0.0
